+++ Source/AMRInterpolator/InterpolationLayout.hpp
@@ -17,7 +17,7 @@ class InterpolationLayout
     std::vector<int> level_idx;
     std::vector<int> box_idx;
 
-    InterpolationLayout(int num_points)
+    InterpolationLayout(size_t num_points)
         : rank(num_points, -1), level_idx(num_points, -1),
           box_idx(num_points, -1)
     {
